[65] '.time' package created, Clock and Timer class added.
The Chrono, Clock and Timer class now all reside inside the newly added '.time' package. Chrono has as SuperClass the Clock Class, as well as the Timer. Timers are new Objects which store a duration of time that can be edited and elapsed when needed (they make the thread sleep for the ammount of time they store).