[96] added JavaDocs, fixed Chrono and Clock and improved Moona's safety.
Both the Clock and Chrono classes received full JavaDoc description, as well as a revamp of the sleep(int nanos) mehtod, which now takes as an argument microseconds instead. The Moona Class received a substantial update in security: now all of its methods will be exclusively available after the .Init() method is called, otherwise a MoonaHandlingException will be thrown.