[151] various changes throughout the framework.
Now all the processes also depend on the Moona.isOn field (which now has a package visibility): their lifetime is dictated by it being true or false. Hence, from now on, processes will be able to run only if Moona gets correctly initialized and they'll collapse if it turns to false. From here on, the only way of turning that bool to false will be to either Collapse() Moona or to get a MoonaHandlingException() thrown.