[170] event framework's development reached a satisfying point.
Events are handled by EventSpaces and EventDaemons (the first always remains active, meanwhile the second one executes an array of events before terminating itself). Events are executed depending on the value of the getMode() method: it returns an Enum field from the EventMode enum stating how it should run. EventMode.ONCE makes the Event execute only once, EventMode.REPEAT makes it iterate over and over until a certain number or even infinitely and the EventMode.UNTIL keeps triggering the Event as long as a certain condition is satisfied.