[212] Moona, Phase and Mirror are now complete.
Those three classes were made completely static. The only operations allowed are static methods invocations, which make up for the new base of the Object Handling System. All of those three classes are connected to the IshMap inside of the Moona Class, which contains all the different Serials; you are able to interact with those elements by using the different methods added to the other classes.