[285] majour Process speed improvements.
Status have been improved: they now can allow for any kind of value to be monitored and the visibility of their methods can be changed. Now Processes do not rely on two separate Statuses, but on just one, which holds a ProcessCondition. For compatibility's sake, .isPaused() and .isRunning() methods were kept as default methods inside the Process Class, since their presence in the code is still rich and changing it was not ideal nor efficient.