[290] Unique detection system done.
Now each Object Handler is able to detect those elements marked with the Unique Annotation. The purpose of this feature is to mark those elements which must not be processed through an handler more than once. Further development might lead up to the user's deliberate choiche on how many times these should be re-processed.