[390] created Reference and Annotated Classes.
The Reference Class extends AbstractReflection<AnnotatedElement> and it contains Nested Classes extending it that each represent a different reflection element (like a method, for example, represented by the Reference.Method Nested Class). The Annotated Class is built the same way as the Reflection Class, but represents an annotated element for each Nested Class.