[432] added Cluster.
Reflection Clusters are groups of Reflections that can all relate to a single target. There are also classes that implement the Cluster Interface, which can be assumed as groups of Reflections with specific properties (the TypeContent Cluster, for example, builds on top of a Type Reference and dumps all the Reflections that can derive from it).