[460] improved .equals() methods and updated Mirror class.
The Mirror class now detects whether or not a Reflection is already contained in it. It does so not by checking IDs but by using the .equals() method, which is an alternate way of comparing Reflections, since it relies on parameters that can identify two Reflections as equal even though their ID isn't the same. This was made in order to prevent duplicate Reflections (meaning Reflections that highlight the same aspects of the same target).