[461] added strictContext.
The strictContext static boolean is a value stored in the Reflection class which states whether or not Reflections must necessarily be added to the Mirror. The Strict Context is a way of stating when certain Reflections are needed to get into the Mirror in case Moona didn't declare auto-reflections as allowed: it comes in handy in cases such as Property Beacons, which generate tons of new Reflections when the .reflect() method gets called. Those Reflection must be included in the Mirror to avoid duplicates.