[489] added Annotation Reference and made Annotated a Flare.
Since Annotated was just a Reflection which evaluated to a List of Annotations I thought that making it a Flare would have been better. On top of that, I also wanted to add a dedicated Reference to Annotations, since just referring at them with a Type Reference would have been risky.