[500] fixed .equals() method for Flares.
Before this fix, a Flare used to check whether or not it was equal to another by using the .equals() method in its superclass (Flare class). This meant, though, that all the Flares which had the same target were considered equal (even if not of the same type), now it works properly. 500th commit, btw ^^.