[517] added trusted boolean to Existing References.
When an Existing Reference's constructor gets called, it detects whether or not the Strict Context is enabled and, if so, evaluates the "trusted" boolean to true (false by default). Doing so avoids useless checks when it comes to evaluation time, since, when the Strict Context is active, no Undefined Reflection can be generated (since it's trusted code, and I (kinda) trust my brain).