[527] improved Reflections' safety.
As soon as the Moona.init() method stops cycling through all the Reflections to initialize them, it locks the autoReflections settings on disabled permanently. This was made in order to prevent massive performance issues when evaluating Reflections after the pre-init phase, since fully evaluating and deriving a Reflection might take a lot of time (probably, the Mirror will cover the loading time with already evaluated Reflections, too).