[602] fixed Objects, Entities and Types.
Now Objects, Types and Entities have different template arguments. When an Object extends the Object or Entity class, the first parameter argument will have to be the Object itself (for a class A extending Object, the correct syntax will be "class A : public Object<A, ...>"). The arguments that come after the first one will be interpreted as the other base-classes of this new Object (I do would not recommend extending other classes in a way different from this).