[608] improved C++ Object's structure.
At their root now there's the Base class, which every Entity (consequently, every Object too) inherit. It's the way of telling actual Object types when using Assertions to check whether or not a class is a Moona Object (that's the main use: a way of "saving" the base type of an Object somewhere).