[612] modified .getTarget() method for some Reference.
Constructor, Method and Field References used to return their reflected value, but now they return their declaring class. It's much easier to think to a declaring class as a Reference's target, since it derived from there (and since I need things to work this way in order to develop an upcoming feature).