[624] totally removed Serials and Naturals.
It was about damn time for me to remove all that system. Object Handling based on Serials and Naturals was... really not useful at all. Eversince I started coding this framework I relied a lot on this system, but I realized it was not ideal at all. Serials got now replaced with MoonaObject(s), which is just a common super interface for all those elements required to stay in Object Handlers.