[637] added MirrorFilters.
MirrorFilters are a way to access Reflections inside the Mirror. Each different type of Reflection features a custom method to get a MirrorFilter which contains all the Reflections of that given type. In order to achieve proper filtering, a new package has been added, the .filter package: it contains different interfaces which highlight traits of every Reflection (for example, if it's a class element, or if it has a name etc...).