[662] added Natures for Processes.
A Process' Nature is an Object which holds its ProcessClock its Status. There is no link whatsoever between this Nature and the previously removed Nature and Natural(s) system, the name was just borrowed. The .nature() method was introduced in the Process interface, replacing the .getClock() and .getStatus() methods (which weren't removed, but just defaulted).