[664] added design root package and Grids.
Grids are classes which needed to manage basic abstract grids, each having a width and a height. There are also ScaledGrids, which can tell in what square of it two given coordinates are. The design package will contain all the elements needed to design abstract phisical patterns (such as grids, points, shapes etc...).