[666] added Bidimensional Elements and GridPoint.
Bidimensional Elements are a bunch of interfaces which role is to define bidimensional elements behaviours in space. The most basic Bidimensional Element was also added, the Point2D. GridPoints, the basic units of Grids, were added as well, using integral precision instead of the decimal one, proper of Point2Ds.