[693] added Bounds.
Bounds are an abstraction of Dimensions. While you're able to initialize a Dimension with negative values (to represent a negative extention on a certain axis), the same can't apply for Bounds. Extracting bounds from a Dimensional means obtaining a Bounds object (the number of dimensions it lives in and its precision are arbitrary) which contains the same parameters as the Dimensional, but stored as natural numbers (positives only).