[696] added .getPosition(), .getBounds() and .getSize() methods.
Positional elements are now able to return their absolute location in space with the .getPosition() method, which returns a Position of the same precision as the considered Positional element. On the other hand, Dimensionals have a similar feature, too: they can return their size in space (allowing extensions on negative axis) with the .getSize() method and they can return their absolute size (the absolute value of their dimensions) with the .getBounds() method.