[702] added interdimensional methods.
Interdimensional methods such as .asX() and .asY() for Positionals and .asWidth() and .asHeight() for Resizables take as an argument an element with one less dimension with respect to the element which calls the method. They change one of the coordinates/dimensions of the Object with the ones of the given argument.